add support for computer isolation / integration
